refactor(consul): create components
- created `ImageSelector` and `DocumentSelector`
  components to improve readability
- updated with `{ status, body }` as there are two
  `uploadData` functions
- delete the `bytesToSize` function as it is a
  previously prepared `formatSize` function
- delete `IMAGE_TYPE_REGEX` from
  `NewProposal.js` and add in `consts`
- fixed typos in `texts` file

SVA-596
