+++ src/components/consul/index.js
@@ -1,4 +1,5 @@
 export * from './detail';
 export * from './list';
+export * from './selectors';
 
 export * from './ConsulWelcome';
