+++ src/components/consul/selectors/index.js
@@ -1,0 +1,2 @@
+export * from './ConsulDocumentSelector';
+export * from './ConsulImageSelector';
